requirements.txt
elasticsearch~=8.6.1
python-dotenv~=0.21.1
requests~=2.28.2